adds email functionality

+++ src/main/java/com/airline/reservation/repository/FlightRepository.java
@@ -5,13 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.List;
 
 @Repository
 public interface FlightRepository extends JpaRepository<Flight, Long> {
     public Flight getByFlightNumber(Integer flightNumber);
     public List<Flight> findByDepartureDate(LocalDate departureDate);
-
-    public Flight getFlightById(Long flightId);
+    public Flight getFlightById(Integer flightId);
 }
